feat: embed query param for usage in a iframe
