+++ frontend/src/i18n/en/index.ts
@@ -20,8 +20,8 @@ export default {
   home: 'Home',
   introduction: 'Introduction',
   resources: 'Resources',
-  fly_to_arctic: 'Fly to Arctic',
-  fly_to_antarctic: 'Fly to Antarctic',
+  fly_to_arctic: 'Arctic',
+  fly_to_antarctic: 'Antarctica',
   base_map: 'Base map',
   instrument_measures: 'Instruments & Measures',
   fundings: 'Fundings',
