--- frontend/src/layouts/SimpleLayout.vue
@@ -1,15 +1,0 @@
-<template>
-  <q-layout view="hHh lpR fFf">
-    <q-header bordered class="bg-white text-grey-10">
-      <app-toolbar no-menu />
-    </q-header>
-
-    <q-page-container>
-      <router-view />
-    </q-page-container>
-  </q-layout>
-</template>
-
-<script setup lang="ts">
-import AppToolbar from 'src/components/AppToolbar.vue';
-</script>
