+++ frontend/src/pages/HomePage.vue
@@ -1,25 +1,41 @@
 <template>
   <q-page :class="mapStore.showGlobe ? 'bg-black' : ''">
-    <div v-if="mapStore.showGlobe">
-      <globe-view />
-    </div>
-    <div v-else>
-      <q-splitter
-        v-model="split"
-        :horizontal="$q.screen.lt.md"
-        style="height: 94vh"
-      >
-        <template v-slot:before>
-          <azimuth-view id="north" projection="EPSG:3995" />
-        </template>
+    <div>
+      <q-btn
+        no-caps
+        color="primary"
+        class="on-left"
+        :icon="mapStore.showGlobe ? 'map' : 'public'"
+        :label="mapStore.showGlobe ? '2D' : '3D'"
+        @click="toggleShowGlobe"
+        style="
+          position: absolute;
+          top: 60px;
+          right: auto;
+          bottom: auto;
+          left: 8px;
+          z-index: 1000;"
+      ></q-btn>
+      <div v-if="mapStore.showGlobe">
+        <globe-view />
+      </div>
+      <div v-else>
+        <q-splitter
+          v-model="split"
+          :horizontal="$q.screen.lt.md"
+        >
+          <template v-slot:before>
+            <azimuth-view id="north" projection="EPSG:3995" :embed="isEmbeded" />
+          </template>
 
-        <template v-slot:separator>
-          <q-avatar color="primary" text-color="white" size="40px" icon="drag_indicator" />
-        </template>
-        <template v-slot:after>
-          <azimuth-view id="south" projection="EPSG:3031" />
-        </template>
-      </q-splitter> 
+          <template v-slot:separator>
+            <q-avatar color="primary" text-color="white" size="40px" icon="drag_indicator" />
+          </template>
+          <template v-slot:after>
+            <azimuth-view id="south" projection="EPSG:3031" :embed="isEmbeded" />
+          </template>
+        </q-splitter> 
+      </div>
     </div>
     
     <simple-dialog 
@@ -38,10 +54,17 @@ import SimpleDialog from 'src/components/SimpleDialog.vue';
 import CampaignView from 'src/components/CampaignView.vue';
 import GlobeView from 'src/components/map/GlobeView.vue';
 import AzimuthView from 'src/components/map/AzimuthView.vue';
+import { Settings } from 'src/stores/settings';
 
+const route = useRoute();
+const router = useRouter();
+const settingsStore = useSettingsStore();
 const mapStore = useMapStore();
+
 const split = ref(50);
 const showCampaign = ref(false);
+
+const isEmbeded = computed(() => route.query.embed !== undefined && route.query.embed !== 'false' && route.query.embed !== '0');
 
 onMounted(() => {
   mapStore.loadCampaigns();
@@ -57,6 +80,14 @@ watch(
 function onShowCampaignUpdated() {
   if (!showCampaign.value) {
     mapStore.selectedCampaign = null;
+  }
+}
+
+function toggleShowGlobe() {
+  mapStore.showGlobe = !mapStore.showGlobe;
+  settingsStore.saveSettings({ show_globe: mapStore.showGlobe } as Settings);
+  if (route.path === '/admin') {
+    router.push('/');
   }
 }
 </script>
